requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
google-generativeai==0.7.2
gunicorn==21.2.0